iOS 10.2

+++ Frameworks/AVFoundation.framework/AVAssetProxyInternal.h
@@ -4,7 +4,7 @@
 
 @interface AVAssetProxyInternal : NSObject {
     AVAssetInspectorLoader * loader;
-    long  makeTracksArrayOnce;
+    long long  makeTracksArrayOnce;
     NSArray * tracks;
 }
