+++ Frameworks/AVFoundation.framework/AVAssetWriterFailedTerminalHelper.h
@@ -10,10 +10,10 @@
 - (void)endSessionAtSourceTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg1;
 - (id)error;
 - (void)finishWriting;
-- (void)finishWritingWithCompletionHandler:(id /* block */)arg1;
+- (void)finishWritingWithCompletionHandler:(id)arg1;
 - (id)initWithConfigurationState:(id)arg1;
 - (id)initWithConfigurationState:(id)arg1 terminalError:(id)arg2;
 - (void)startSessionAtSourceTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg1;
-- (int)status;
+- (long long)status;
 
 @end
