+++ Frameworks/AVFoundation.framework/AVCaptureStillImageRequest.h
@@ -3,35 +3,35 @@
  */
 
 @interface AVCaptureStillImageRequest : NSObject {
-    id /* block */  _bracketedCaptureCompletionBlock;
+    id  _bracketedCaptureCompletionBlock;
     AVCaptureBracketedStillImageSettings * _bracketedSettings;
-    id /* block */  _iosurfaceCompletionBlock;
-    id /* block */  _sbufCompletionBlock;
+    id  _iosurfaceCompletionBlock;
+    id  _sbufCompletionBlock;
     long long  _settingsID;
-    unsigned long  _shutterSoundID;
+    unsigned int  _shutterSoundID;
 }
 
-@property (copy) id /* block */ bracketedCaptureCompletionBlock;
+@property (copy) id bracketedCaptureCompletionBlock;
 @property (retain) AVCaptureBracketedStillImageSettings *bracketedSettings;
-@property (copy) id /* block */ iosurfaceCompletionBlock;
-@property (copy) id /* block */ sbufCompletionBlock;
+@property (copy) id iosurfaceCompletionBlock;
+@property (copy) id sbufCompletionBlock;
 @property long long settingsID;
-@property unsigned long shutterSoundID;
+@property unsigned int shutterSoundID;
 
 + (id)request;
 
-- (id /* block */)bracketedCaptureCompletionBlock;
+- (id)bracketedCaptureCompletionBlock;
 - (id)bracketedSettings;
 - (void)dealloc;
-- (id /* block */)iosurfaceCompletionBlock;
-- (id /* block */)sbufCompletionBlock;
-- (void)setBracketedCaptureCompletionBlock:(id /* block */)arg1;
+- (id)iosurfaceCompletionBlock;
+- (id)sbufCompletionBlock;
+- (void)setBracketedCaptureCompletionBlock:(id)arg1;
 - (void)setBracketedSettings:(id)arg1;
-- (void)setIosurfaceCompletionBlock:(id /* block */)arg1;
-- (void)setSbufCompletionBlock:(id /* block */)arg1;
+- (void)setIosurfaceCompletionBlock:(id)arg1;
+- (void)setSbufCompletionBlock:(id)arg1;
 - (void)setSettingsID:(long long)arg1;
-- (void)setShutterSoundID:(unsigned long)arg1;
+- (void)setShutterSoundID:(unsigned int)arg1;
 - (long long)settingsID;
-- (unsigned long)shutterSoundID;
+- (unsigned int)shutterSoundID;
 
 @end
