+++ Frameworks/AVFoundation.framework/AVExternalDeviceInternal.h
@@ -7,8 +7,8 @@
     NSData * _MFiCertSerial;
     <AVExternalDeviceDelegate> * _delegate;
     struct OpaqueFigEndpoint { } * _figEndpoint;
-    long  _makeHIDsOnlyOnce;
-    long  _makeIconsOnlyOnce;
+    long long  _makeHIDsOnlyOnce;
+    long long  _makeIconsOnlyOnce;
     NSArray * _oemIcons;
     NSObject<OS_dispatch_queue> * _queue;
     NSArray * _screenInfo;
