+++ Frameworks/AVFoundation.framework/AVMetadataItemInternal.h
@@ -15,7 +15,7 @@
     NSString * extendedLanguageTag;
     NSDictionary * extras;
     NSString * identifier;
-    long  itemIndex;
+    long long  itemIndex;
     <NSObject><NSCopying> * key;
     NSString * keySpace;
     NSString * languageCode;
