+++ Frameworks/AddressBookUI.framework/ABActionSheet.h
@@ -12,19 +12,19 @@
 
 @property (nonatomic) void*ab_context;
 
-- (int)_addButtonWithTitle:(id)arg1;
+- (long long)_addButtonWithTitle:(id)arg1;
 - (id)_addButtonWithTitle:(id)arg1 label:(id)arg2;
-- (id)ab_addButtonWithTitle:(id)arg1 label:(id)arg2 tag:(int)arg3;
-- (id)ab_addButtonWithTitle:(id)arg1 label:(id)arg2 tag:(int)arg3 tag2:(int)arg4;
-- (id)ab_addButtonWithTitle:(id)arg1 label:(id)arg2 tag:(int)arg3 tag2:(int)arg4 tag3:(const void*)arg5;
-- (id)ab_addButtonWithTitle:(id)arg1 label:(id)arg2 tag:(int)arg3 tag2:(int)arg4 tag3:(const void*)arg5 tag4:(const void*)arg6;
-- (int)ab_addButtonWithTitle:(id)arg1 tag:(int)arg2;
-- (int)ab_addButtonWithTitle:(id)arg1 tag:(int)arg2 tag2:(int)arg3;
+- (id)ab_addButtonWithTitle:(id)arg1 label:(id)arg2 tag:(long long)arg3;
+- (id)ab_addButtonWithTitle:(id)arg1 label:(id)arg2 tag:(long long)arg3 tag2:(long long)arg4;
+- (id)ab_addButtonWithTitle:(id)arg1 label:(id)arg2 tag:(long long)arg3 tag2:(long long)arg4 tag3:(const void*)arg5;
+- (id)ab_addButtonWithTitle:(id)arg1 label:(id)arg2 tag:(long long)arg3 tag2:(long long)arg4 tag3:(const void*)arg5 tag4:(const void*)arg6;
+- (long long)ab_addButtonWithTitle:(id)arg1 tag:(long long)arg2;
+- (long long)ab_addButtonWithTitle:(id)arg1 tag:(long long)arg2 tag2:(long long)arg3;
 - (void*)ab_context;
-- (int)ab_tag2AtIndex:(int)arg1;
-- (const void*)ab_tag3AtIndex:(int)arg1;
-- (const void*)ab_tag4AtIndex:(int)arg1;
-- (int)ab_tagAtIndex:(int)arg1;
+- (long long)ab_tag2AtIndex:(long long)arg1;
+- (const void*)ab_tag3AtIndex:(long long)arg1;
+- (const void*)ab_tag4AtIndex:(long long)arg1;
+- (long long)ab_tagAtIndex:(long long)arg1;
 - (void)dealloc;
 - (id)init;
 - (id)initWithTitle:(id)arg1 delegate:(id)arg2 cancelButtonTitle:(id)arg3 destructiveButtonTitle:(id)arg4 otherButtonTitles:(id)arg5;
