+++ Frameworks/AddressBookUI.framework/ABActionSheetDelegate.h
@@ -9,10 +9,10 @@
 
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (readonly) Class superclass;
 
-- (void)actionSheet:(id)arg1 didDismissWithButtonIndex:(int)arg2;
+- (void)actionSheet:(id)arg1 didDismissWithButtonIndex:(long long)arg2;
 - (id)initWithActionSheetTarget:(id)arg1 action:(SEL)arg2;
 
 @end
