+++ Frameworks/AddressBookUI.framework/ABActionsController.h
@@ -18,17 +18,17 @@
 + (int)defaultActionForProperty:(int)arg1;
 + (id)newActionsControllerForProperty:(int)arg1;
 
-- (int)_actionAtIndex:(int)arg1 forProperty:(int)arg2;
+- (int)_actionAtIndex:(long long)arg1 forProperty:(int)arg2;
 - (struct __CFArray { }*)_actionsForProperty:(int)arg1;
 - (void)_performAction:(int)arg1 forProperty:(int)arg2;
-- (int)actionsCountForProperty:(int)arg1;
+- (long long)actionsCountForProperty:(int)arg1;
 - (void*)context;
 - (void)dealloc;
 - (int)defaultAction;
 - (id)init;
 - (id)initWithProperty:(int)arg1;
 - (void)performAction:(int)arg1;
-- (void)performActionAtIndex:(int)arg1 forProperty:(int)arg2;
+- (void)performActionAtIndex:(long long)arg1 forProperty:(int)arg2;
 - (void)performDefaultAction;
 - (id)person;
 - (void)resetPrimaryPropertyActions;
@@ -38,7 +38,7 @@
 - (void)setStringValue:(id)arg1;
 - (void)setValue:(id)arg1;
 - (id)stringValue;
-- (id)titleAtIndex:(int)arg1 forProperty:(int)arg2;
+- (id)titleAtIndex:(long long)arg1 forProperty:(int)arg2;
 - (id)titleForAction:(int)arg1;
 - (id)urlForAction:(int)arg1 forProperty:(int)arg2;
 - (id)value;
