+++ Frameworks/AddressBookUI.framework/ABPersonOverlayingContainerView.h
@@ -5,12 +5,12 @@
 @interface ABPersonOverlayingContainerView : UIScrollView <ABPersonTabsScrollView> {
     UIView * _abContentView;
     UIView * _backgroundView;
-    float  _minimumBottomInset;
+    double  _minimumBottomInset;
     struct UIEdgeInsets { 
-        float top; 
-        float left; 
-        float bottom; 
-        float right; 
+        double top; 
+        double left; 
+        double bottom; 
+        double right; 
     }  _savedContentInset;
     <ABPersonTabsLayoutManager> * _tabsLayoutManager;
 }
@@ -19,8 +19,8 @@
 @property (nonatomic, retain) UIView *contentView;
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
-@property (nonatomic) float minimumBottomInset;
+@property (readonly) unsigned long long hash;
+@property (nonatomic) double minimumBottomInset;
 @property (readonly) Class superclass;
 @property (nonatomic, retain) <ABPersonTabsLayoutManager> *tabsLayoutManager;
 
@@ -29,13 +29,13 @@
 - (id)contentView;
 - (void)dealloc;
 - (void)layoutSubviews;
-- (float)minimumBottomInset;
+- (double)minimumBottomInset;
 - (void)setBackgroundView:(id)arg1;
-- (void)setContentInset:(struct UIEdgeInsets { float x1; float x2; float x3; float x4; })arg1;
-- (void)setContentOffset:(struct CGPoint { float x1; float x2; })arg1;
+- (void)setContentInset:(struct UIEdgeInsets { double x1; double x2; double x3; double x4; })arg1;
+- (void)setContentOffset:(struct CGPoint { double x1; double x2; })arg1;
 - (void)setContentView:(id)arg1;
-- (void)setFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
-- (void)setMinimumBottomInset:(float)arg1;
+- (void)setFrame:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
+- (void)setMinimumBottomInset:(double)arg1;
 - (void)setTabsLayoutManager:(id)arg1;
 - (id)tabsLayoutManager;
