+++ Frameworks/CFNetwork.framework/_NSCFTranslatedFileURLProtocol_PIMPL_3.h
@@ -4,6 +4,6 @@
 
 @interface _NSCFTranslatedFileURLProtocol_PIMPL_3 : _NSCFTranslatedFileURLProtocol
 
-+ (const struct CFURLProtocolImplementation { int x1; void *x2; int (*x3)(); int (*x4)(); int (*x5)(); int (*x6)(); int (*x7)(); int (*x8)(); int (*x9)(); }*)pimpl;
++ (const struct CFURLProtocolImplementation { long long x1; void *x2; int (*x3)(); int (*x4)(); int (*x5)(); int (*x6)(); int (*x7)(); int (*x8)(); int (*x9)(); }*)pimpl;
 
 @end
