+++ Frameworks/Contacts.framework/CNIndexRequestHandler.h
@@ -11,7 +11,7 @@
 
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
-@property (readonly) unsigned int hash;
+@property (readonly) unsigned long long hash;
 @property (nonatomic, readonly) CSSearchableIndex *index;
 @property (readonly) Class superclass;
 
@@ -28,8 +28,8 @@
 - (id)index;
 - (id)init;
 - (id)initWithAddressBook:(void*)arg1;
-- (void)searchableIndex:(id)arg1 reindexAllSearchableItemsWithAcknowledgementHandler:(id /* block */)arg2;
-- (void)searchableIndex:(id)arg1 reindexSearchableItemsWithIdentifiers:(id)arg2 acknowledgementHandler:(id /* block */)arg3;
+- (void)searchableIndex:(id)arg1 reindexAllSearchableItemsWithAcknowledgementHandler:(id)arg2;
+- (void)searchableIndex:(id)arg1 reindexSearchableItemsWithIdentifiers:(id)arg2 acknowledgementHandler:(id)arg3;
 - (id)searchableItemForContactIdentifier:(id)arg1;
 - (id)searchableItemForPerson:(void*)arg1;
